requirements.txt
pydicom
opencv-python
numpy
pillow
pandas
tensorflow
scikit-image
matplotlib